% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chickens.R
\name{getNumberOfChickens}
\alias{getNumberOfChickens}
\title{Get number of chickens at given time}
\usage{
getNumberOfChickens(state)
}
\arguments{
\item{state}{State vector}
}
\value{
Number of chickens (total) in the given state
}
\description{
Get number of chickens at given time
}
