% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chickens.R
\name{getNumberOfChickensAtAllTimes}
\alias{getNumberOfChickensAtAllTimes}
\title{Get the number of chickens at all times}
\usage{
getNumberOfChickensAtAllTimes(realisation)
}
\arguments{
\item{realisation}{Realisation from the Chickens model}
}
\value{
Vector consisting of the number of chickens at each time
}
\description{
Get the number of chickens at all times
}
\examples{
df <- runChickensModel(parameter_list = p, betas=betas)
df$realisation$total_chickens <- getNumberOfChickensAtAllTimes(df$realisation)
}
