% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chickens.R
\name{runChickensModel}
\alias{runChickensModel}
\title{Run Chicken Model}
\usage{
runChickensModel(parameter_list, betas = matrix(), dt = 1,
  max_time = 1000, solver_type = "stochastic")
}
\arguments{
\item{parameter_list}{A list of parameters for this realisation. Needs the following structure:
\itemize{
\item{\code{patchname}: One of "Es","Ns","Bs" or "Sc"}
  \itemize{
     \item{\code{x0}: Initial condition}
     \item{\code{delta}: Numeric vector of length 5}
     \item{\code{y}: Numeric in [0, 1]}
     \item{\code{x}: Numeric in [0, 1]}
     \item{\code{alpha}: List containing 3 elements, "lG","He" and "Rs"}
     \item{\code{sigma}:}
     \item{\code{gamma}:}
     \item{\code{w}:}
     \item{\code{n_egg}:}
     \item{\code{K}: Carrying capacity (Sc system only)}
     \item{\code{q}:}
  }
}
Repeat for each possible patch.}

\item{betas}{Matrix of within and between patch transmission (row names are required)}

\item{dt}{Time spacing of outputs (NOT solving points)}

\item{max_time}{Maximum time for simulation}

\item{solver_type}{Either "stochastic" or "deterministic"}
}
\value{
A list containing two elements: \code{realisation}, contains the realisation and \code{parameters} contains the parameters
}
\description{
Runs a single realisation of the chickens model
}
\examples{
betas <- matrix(1.5, dimnames=list(c("Es")))
x0 <- list("E"=100, "Ch.S"=50, "He.S"=50, "He.I"=10)
p_sub <- list("x0"=x0)
p <- list("Es"=p_sub)
df <- runChickensModel(parameter_list = p, betas=betas)

}
